moved new functions to its own file and added test files
