+++ testdata/param_structs/inner/inner.go
@@ -1,0 +1,10 @@
+package inner
+
+import (
+	structs "github.com/yalochat/swag/v2/testdata/param_structs"
+)
+
+type InnerStruct struct {
+	AwesomeField string `json:"awesomeField"`
+	FormModelExample *structs.FormModel `json:"formModelExample"`
+}
